+++ AppYogaDemo/app/src/main/java/com/example/appyogademo/ViewModels/YogaClassViewModel.java
@@ -1,6 +1,8 @@
 package com.example.appyogademo.ViewModels;
 
+import android.app.AlertDialog;
 import android.content.Context;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.LiveData;
@@ -8,6 +10,7 @@ import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.example.appyogademo.Models.YogaClass;
+import com.example.appyogademo.Repository.Result;
 import com.example.appyogademo.Repository.YogaClassRepository;
 
 import java.util.List;
@@ -15,8 +18,9 @@ import java.util.List;
 public class YogaClassViewModel extends ViewModel {
     private final YogaClassRepository repository;
     private final LiveData<List<YogaClass>> yogaClasses;
-
+    private final Context context;
     public YogaClassViewModel(Context context) {
+        this.context = context;
         repository = new YogaClassRepository(context);
         yogaClasses = repository.getAllYogaClasses();
     }
@@ -25,32 +29,52 @@ public class YogaClassViewModel extends ViewModel {
         return yogaClasses;
     }
 
-    public void insert(YogaClass yogaClass) {
-        repository.insert(yogaClass);
+    public Boolean insert(YogaClass yogaClass) {
+        try {
+            Result result = repository.insert(yogaClass);
+            if(result.isStatus()) {
+                Toast.makeText(context, result.getMessage(), Toast.LENGTH_SHORT).show();
+                return true;
+            } else {
+                Toast.makeText(context, result.getMessage(), Toast.LENGTH_SHORT).show();
+                return false;
+            }
+        } catch (Exception e) {
+            Toast.makeText(context, e.toString(), Toast.LENGTH_SHORT).show();
+            return false;
+        }
     }
 
     public void update(YogaClass yogaClass) {
-        repository.update(yogaClass);
+
+        try {
+            Result result = repository.update(yogaClass);
+            if(result.isStatus()) {
+                Toast.makeText(context, result.getMessage(), Toast.LENGTH_SHORT).show();
+            } else {
+                Toast.makeText(context, result.getMessage(), Toast.LENGTH_SHORT).show();
+            }
+        } catch (Exception e) {
+            Toast.makeText(context, e.toString(), Toast.LENGTH_SHORT).show();
+        }
     }
 
     public void delete(YogaClass yogaClass) {
-        repository.delete(yogaClass);
-    }
-
-    public static class Factory extends ViewModelProvider.NewInstanceFactory {
-        private final Context context;
-
-        public Factory(Context context) {
-            this.context = context;
-        }
-
-        @NonNull
-        @Override
-        public <T extends ViewModel> T create(Class<T> modelClass) {
-            if (modelClass.isAssignableFrom(YogaClassViewModel.class)) {
-                return (T) new YogaClassViewModel(context);
-            }
-            throw new IllegalArgumentException("Unknown ViewModel class");
+        try {
+            new AlertDialog.Builder(context)
+                    .setTitle("Xóa khóa học")
+                    .setMessage("Bạn có chắc chắn muốn xóa class này không?")
+                    .setPositiveButton("Có", (dialog, which) -> {
+                        repository.delete(yogaClass);
+                        Toast.makeText(context, "Đã xóa khóa học", Toast.LENGTH_SHORT).show();
+                    })
+                    .setNegativeButton("Không", (dialog, which) -> {
+                        dialog.dismiss();
+                    })
+                    .show();
+        } catch (Exception e) {
+            Toast.makeText(context, e.toString(), Toast.LENGTH_SHORT).show();
         }
     }
+
 }
